extract ButtonViewHolder from Button model
