+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Button.java
@@ -1,27 +1,16 @@
 package org.cru.godtools.tract.model;
 
-import android.content.Context;
-import android.content.res.ColorStateList;
 import android.net.Uri;
 import android.support.annotation.ColorInt;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.annotation.UiThread;
 import android.support.annotation.WorkerThread;
-import android.support.v4.view.ViewCompat;
-import android.view.ViewGroup;
-import android.widget.TextView;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableSet;
 
 import org.ccci.gto.android.common.util.XmlPullParserUtils;
-import org.cru.godtools.analytics.AnalyticsService;
 import org.cru.godtools.base.model.Event;
-import org.cru.godtools.base.ui.util.WebUrlLauncher;
-import org.cru.godtools.tract.R;
-import org.cru.godtools.tract.R2;
-import org.cru.godtools.tract.model.AnalyticsEvent.Trigger;
 import org.cru.godtools.tract.model.Text.Align;
 import org.jetbrains.annotations.Contract;
 import org.xmlpull.v1.XmlPullParser;
@@ -30,9 +19,6 @@ import org.xmlpull.v1.XmlPullParserException;
 import java.io.IOException;
 import java.util.Collection;
 import java.util.Set;
-
-import butterknife.BindView;
-import butterknife.OnClick;
 
 import static org.cru.godtools.tract.Constants.XMLNS_ANALYTICS;
 import static org.cru.godtools.tract.Constants.XMLNS_CONTENT;
@@ -48,7 +34,7 @@ public final class Button extends Content implements Styles {
 
     private static final Align DEFAULT_TEXT_ALIGN = Align.CENTER;
 
-    private enum Type {
+    enum Type {
         EVENT, URL, UNKNOWN;
 
         static final Type DEFAULT = UNKNOWN;
@@ -156,47 +142,5 @@ public final class Button extends Content implements Styles {
         mType = Type.parse(parser.getAttributeValue(null, XML_TYPE), mType);
         mEvents = parseEvents(parser, XML_EVENTS);
         mUrl = Utils.parseUrl(parser, XML_URL, null);
-    }
-
-    @UiThread
-    static final class ButtonViewHolder extends BaseViewHolder<Button> {
-        @BindView(R2.id.button)
-        TextView mButton;
-
-        ButtonViewHolder(@NonNull final ViewGroup parent, @Nullable final BaseViewHolder parentViewHolder) {
-            super(Button.class, parent, R.layout.tract_content_button, parentViewHolder);
-        }
-
-        /* BEGIN lifecycle */
-
-        @Override
-        void onBind() {
-            super.onBind();
-            final Text text = mModel != null ? mModel.mText : null;
-            Text.bind(text, mButton);
-            ViewCompat.setBackgroundTintList(mButton, ColorStateList.valueOf(getButtonColor(mModel)));
-        }
-
-        /* END lifecycle */
-
-        @OnClick(R2.id.button)
-        void click() {
-            if (mModel != null) {
-                triggerAnalyticsEvents(mModel.mAnalyticsEvents, Trigger.SELECTED, Trigger.DEFAULT);
-
-                switch (mModel.mType) {
-                    case URL:
-                        if (mModel.mUrl != null) {
-                            final Context context = mRoot.getContext();
-                            AnalyticsService.getInstance(context).onTrackExitUrl(mModel.mUrl);
-                            WebUrlLauncher.openUrl(context, mModel.mUrl);
-                        }
-                        break;
-                    case EVENT:
-                        sendEvents(mModel.mEvents);
-                        break;
-                }
-            }
-        }
     }
 }
